键盘基本实现

+++ YueAi/Controllers/Search/Controller/SearchViewController.m
@@ -174,7 +174,7 @@
 #pragma mark - UITableViewDelegate
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSLog(@"tableView select row at indexpath.row = %ld", indexPath.row);
+    NSLog(@"tableView select row at indexpath.row = %ld", (long)indexPath.row);
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
